vendor: update version system
The version field should only contain the version, we can specify the
preview with the API level for now.

New system for release tags:

Official Patch Update (OPU) will be used only for OTA's, most likely
delta updates.
Official Factory Update (OFU) will be the default increment for full
zips.

We can use the release tag in the zip instead of the generic date, since
the release tag contains the date anyway. We'll just have to keep it
updated
on our own.

Change-Id: I5ad854b94fb8a9eb14e682da2d96579546b035c0